requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
sentencepiece>=0.1.99
accelerate>=0.24.0
bitsandbytes>=0.41.0

# PDF/DOC processing
pdfplumber>=0.10.0
python-docx>=0.8.11
PyPDF2>=3.0.0

# Text processing
nltk>=3.8.0
spacy>=3.7.0
scikit-learn>=1.3.0

# Web/API (optional)
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0

# Utilities
pydantic>=2.5.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.66.0